no_repetDownlaod

+++ langchain/run_gradio.py
@@ -2,18 +2,30 @@ import os
 print('工作路径',os.getcwd())
 print('当前目录',os.path.abspath(os.curdir))
 os.system("python -m pip install --upgrade pip;pip install modelscope==1.9.5;pip install transformers==4.35.2;pip install streamlit==1.24.0;pip install sentencepiece==0.1.99;pip install accelerate==0.24.1;pip install chromadb==0.4.15;pip install sentence-transformers==2.2.2;pip install unstructured==0.10.30;pip install markdown==3.3.7")
-os.system("pip install -U openxlab")
-from openxlab.model import download
-download(model_repo='OpenLMLab/InternLM-chat-7b-8k',output='/home/xlab-app-center/InternLM-chat-7b-8k')
+
+
 os.system("pip install langchain==0.0.292")
 os.system("git clone https://gitee.com/yzy0612/nltk_data.git  --branch gh-pages;cd nltk_data;mv packages/*  ./;cd tokenizers;unzip punkt.zip;cd ../taggers;unzip averaged_perceptron_tagger.zip")
 
-# 设置环境变量
-os.environ['HF_ENDPOINT'] = 'https://hf-mirror.com'
-# 下载模型
-os.system('pip install -U huggingface_hub; huggingface-cli download --resume-download sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2 --local-dir /home/xlab-app-center/sentence-transformer')
-os.system("pip install -r requirements.txt")
-os.system("pip install opencv-python;pip install pytesseract;pip install python-docx;pip install -U pypdf;")#pip install chromadb==0.3.29;
+
+#download model
+if os.path.exists('/home/xlab-app-center/InternLM-chat-7b-8k'):
+    continue
+else:
+    os.system("pip install -U openxlab")
+    from openxlab.model import download
+    download(model_repo='OpenLMLab/InternLM-chat-7b-8k',output='/home/xlab-app-center/InternLM-chat-7b-8k')
+if os.path.exists('/home/xlab-app-center/sentence-transformer'):
+    continue
+else:
+    # 设置环境变量
+    os.environ['HF_ENDPOINT'] = 'https://hf-mirror.com'
+    # 下载模型
+    os.system('pip install -U huggingface_hub; huggingface-cli download --resume-download sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2 --local-dir /home/xlab-app-center/sentence-transformer')
+
+
+os.system("pip install -r langchain/requirements.txt")
+os.system("pip install chromadb==0.3.29;pip install opencv-python;pip install pytesseract;pip install python-docx;pip install -U pypdf;")#pip install chromadb==0.3.29;
 os.system("python langchain/create_db.py")
 # 导入必要的库
 print(os.getcwd())
